% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnmf.R
\name{ebnmf}
\alias{ebnmf}
\title{Smoothed Poisson Topic Model}
\usage{
ebnmf(
  X,
  K,
  lib_size = NULL,
  init = "fasttopics",
  maxiter = 50,
  maxiter_init = 100,
  tol = 0.001,
  ebpm.fn = c(ebpm::ebpm_gamma, smashrgen::ebpm_pois_sgp),
  fix_F = FALSE,
  smooth_F = TRUE,
  smooth_control = list(maxiter = 5, m = 100),
  warm_start = TRUE,
  printevery = 10,
  verbose = TRUE,
  convergence_criteria = "ELBO",
  n_cores = 1
)
}
\arguments{
\item{X}{count matrix}

\item{K}{number of factors/ranks}

\item{lib_size}{library size for RNA seq data}

\item{init}{initialization methods, default is 'fasttopics'; or provide init as a list with L_init, and F_init.}

\item{maxiter, maxiter_init}{maximum iterations}

\item{tol}{stop criteria}

\item{ebpm.fn}{specify functions to use for solving the poisson subproblems}

\item{fix_F}{if TRUE, F will not be updated and will be fixed at the input value in init.}

\item{smooth_F}{whether smooth l or f, must match the functions in ebpm.fn}

\item{smooth_control}{a list. ebnmf_smooth_control_default() gives default settings.}

\item{convergence_criteria}{'mKLabs', or 'ELBO'}

\item{n_cores}{number of cores for parallel update}
}
\value{
EL,EF: posterior of loadings and factors
}
\description{
This function fits Poisson Topic Model with smooth Loading or Factors
}
\examples{
set.seed(123)
n = 120
p = 256
K= 3
L = matrix(0, nrow=n, ncol=K)
FF = matrix(0, nrow=K, ncol=p)
L[1:(n/3),1] = 1
L[((n/3)+1):(2*n/3),2] = 1
L[((2*n/3)+1):n,3] = 1
L = L + matrix(runif(n*K,0,0.5),nrow=n)
FF[1,1:(p/3)] = 1+10
FF[2,((p/3)+1):(2*p/3)] = 1+10
FF[3,((2*p/3)+1):p] = 1+10
lambda = L \%*\% FF
X = matrix(rpois(n=length(lambda),lambda),nrow=n)
image(X)
}
