% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnmf_init.R
\name{ebnmf_init}
\alias{ebnmf_init}
\title{initialize the ebpmf identity model}
\usage{
ebnmf_init(X, K, init, lib_size, maxiter_init = 50)
}
\arguments{
\item{X}{input data matrix}

\item{K}{number of topics}

\item{init}{init methods, or a list of init L and F}
}
\description{
initialize the ebpmf identity model
}
