% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnmf.R
\name{ebnmf_smooth_control_default}
\alias{ebnmf_smooth_control_default}
\title{Default parameters of smooth split}
\usage{
ebnmf_smooth_control_default()
}
\description{
#'@title Default parameters of ebpm
#'@export
ebpm_control_default = function(){
  list(pi0 = 'estimate',
       g_init = NULL,
       fix_g = FALSE,
       control =  NULL)
}
}
