% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnmf.R
\name{ebnmf_update_rank1}
\alias{ebnmf_update_rank1}
\title{rank 1 update of the model}
\usage{
ebnmf_update_rank1(
  l_seq,
  f_seq,
  k,
  ebpm.fn.l,
  ebpm.fn.f,
  res,
  fix_F,
  smooth_F,
  ebpm_control,
  warm_start,
  over_dispersion
)
}
\description{
rank 1 update of the model
}
